refactor(llm): split chat command into commands/chat.tsx subplugin
Move the entire ;chat command (and its supporting helpers — splitContent,
RANDOM_ERROR_MSG, ENABLE_SEARCH_KEYWORDS, shouldEnableSearch) into a
BasePlugin subclass loaded via ctx.plugin(). Like commands/admin.tsx, it
reads everything from ctx.llm so it carries no state of its own.

Drop _adjustDpskV4Prompt: dead code, no call site survives the
extraction. If DeepSeek V4 ever needs the system-prompt-into-user-message
hack again, history shows where to put it back.

PluginLLM is now a pure composition root:
- declares config + db schema + service fields
- registers internal LLM tools
- wires ready hooks (catalog rebuild, image-cache cleanup)
- loads two subplugins (admin commands + chat command)

Line counts:
- index.tsx: 1578 → 342 (−78%)
- new commands/chat.tsx: ~510
- new commands/admin.tsx: ~250

Each subplugin / service file is independently testable and replaceable.

Via Claude Code
